Refactor: remove AES as public API
