+++ src/transaction/__tests__/transaction-erc20.test.ts
@@ -46,7 +46,7 @@ import { RailgunSmartWalletContract } from '../../contracts/railgun-smart-wallet
 import { BoundParamsStruct } from '../../abi/typechain/RailgunSmartWallet';
 import { PollingJsonRpcProvider } from '../../provider/polling-json-rpc-provider';
 import { UTXOMerkletree } from '../../merkletree/utxo-merkletree';
-import { AES } from '../../utils';
+import { AES } from '../../utils/encryption/aes';
 import { POI } from '../../poi/poi';
 import { PoseidonMerkleVerifier } from '../../abi/typechain';
 import { addChainSupportsV3 } from '../../chain/chain';
