+++ src/utils/index.ts
@@ -3,5 +3,4 @@ export { ByteLength, ByteUtils } from './bytes';
 export * from './encryption/ciphertext';
 export { convertTransactionStructToCommitmentSummary } from './commitment';
 export * from './ecies';
-export * from './encryption/aes';
 export { getPublicViewingKey, verifyED25519 } from './keys-utils';
